Fixed first round of review comments
